controller/requirements.txt
Flask==2.0.3
line-bot-sdk==1.20.0
configparser==5.0.2
